feat: Implement a new Bubble Tea-based file manager with directory listing, file operations, and conflict resolution.
