+++ README.md
@@ -1,2 +1,18 @@
-# double
-2 pane file manager in tui that can open files
+# Double Manager
+
+A minimalistic two-pane TUI file manager in Go.
+
+## Running the Application
+
+Since the code is split into multiple files, you need to run the package, not just `main.go`.
+
+```bash
+go run .
+```
+
+Or build it first:
+
+```bash
+go build .
+./double
+```
